refactor: class node, router, rpc_stack
